CustomPathwayPipeline/pyproject.toml
[project]
name = "unstructured_env"
version = "0.1.0"
requires-python = ">=3.9,<=3.12.1"

dependencies = [
    "unstructured[all-docs]",
    "langchain",
    "langchain-community",
    "onnxruntime-gpu",
    "pytesseract",
    "pathway[xpack-llm]",
    "pathway[xpack-llm-docs]",
    "sentence-transformers",
    "bitsandbytes",
    "flask",
    "nltk",
    "keybert",
    "regex",
    "openai",
    "redis",
    "flask-session"
]